[fix] Fix comments
